separating the content into different files and cleaning puts and gets
