+++ lib/game_board.rb
@@ -16,14 +16,4 @@ class Board
       [2, 4, 6]
     ]
   end
-
-  def display_board
-    puts " \t\t \n".yellow
-    puts " \t\t#{@board[0]}|#{@board[1]}|#{@board[2]}"
-    puts " \t\t----- ".yellow
-    puts " \t\t#{@board[3]}|#{@board[4]}|#{@board[5]}"
-    puts " \t\t----- ".yellow
-    puts " \t\t#{@board[6]}|#{@board[7]}|#{@board[8]}"
-    puts " \t\t----- \n".yellow
-  end
 end
