+++ lib/user.rb
@@ -6,8 +6,4 @@ class Player
     @name = name
     @value = value
   end
-
-  def show_players_info
-    puts "player #{@name.green} choose #{@value.yellow}"
-  end
 end
